pyproject.toml
[project]
name = "timer-mute"
version = "0.1.0"
description = "timer mute"
authors = [
    { name = "shift" }
]
dependencies = [
    "coverage>=7.5.3",
    "freezegun>=1.5.1",
    "isort>=5.13.2",
    "mock>=5.1.0",
    "orjson>=3.10.5",
    "pycodestyle>=2.11.1",
    "pysimplegui==4.60.5",
    "ruff>=0.4.8",
    "sqlalchemy>=2.0.30",
    "twitter-api-client>=0.10.22",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
line_length = 119
multi_line_output = 9

[tool.ruff]
exclude = [
    "__pycache__",
    ".coveragerc",
    ".git",
    ".venv",
    "venv",
]
line-length = 119

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "F"]
ignore = ["F401", "F541", "F841"]
preview = true

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/timer_mute"]
